submits/400422047/final/final/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
django-rest-swagger==2.2.0
Django==4.0.5
djangorestframework==3.13.1
drf-yasg==1.20.0
fonttools==4.33.3
idna==3.3
imbalanced-learn==0.9.1
imblearn==0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
Khayyam==3.0.17
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.23.0
openapi-codec==1.3.2
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
ruamel.yaml.clib==0.2.6
ruamel.yaml==0.17.21
scikit-learn==1.1.1
scipy==1.8.1
setuptools==57.0.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
uritemplate==4.1.1
urllib3==1.26.9
wheel==0.36.2
